+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fiori-elements-writer
 
+## 3.0.18
+
+### Patch Changes
+
+- @sap-ux/annotation-generator@1.0.8
+- @sap-ux/fe-fpm-writer@1.0.8
+
+## 3.0.17
+
+### Patch Changes
+
+- Updated dependencies [c9b0659]
+    - @sap-ux/ui5-test-writer@1.1.1
+
 ## 3.0.16
 
 ### Patch Changes