+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 3.0.16
+
+### Patch Changes
+
+- Updated dependencies [c9b0659]
+    - @sap-ux/ui5-test-writer@1.1.1
+
 ## 3.0.15
 
 ### Patch Changes